+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ###[3.0.0-alpha4] - 2016-11-02
 ###Fixed
 - Event Dispatcher Bridge use EventDispatcherInterface instead of EventDispatcher
-- Convert this bundle into an universal package, respecting the PSR11 draft and container-interop/definition-interop 
+- Convert this bundle into an universal package, respecting the PSR11 draft and container-interop/service-provider 
   to be use with any framework
 - Update documentation